backend/requirements.txt
azure-core==1.34.0
azure-storage-blob==12.25.1
blinker==1.9.0
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
cryptography==45.0.3
Flask==3.1.1
flask-cors==6.0.0
idna==3.10
isodate==0.7.2
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
pycparser==2.22
python-dotenv==1.1.0
requests==2.32.3
six==1.17.0
typing_extensions==4.14.0
urllib3==2.4.0
Werkzeug==3.1.3
